+++ ricaun.Nuke.PackageBuilder/Components/Navisworks/IHazNavisworksPackageBuilder.cs
@@ -1,0 +1,13 @@
+﻿using Nuke.Common;
+
+namespace ricaun.Nuke.Components
+{
+    /// <summary>
+    /// Represents a Navisworks package builder component for Nuke build automation.
+    /// Inherits from <see cref="IHazPackageBuilderProject"/> and <see cref="INukeBuild"/>.
+    /// </summary>
+    public interface IHazNavisworksPackageBuilder : IHazPackageBuilderProject, INukeBuild
+    {
+
+    }
+}