+++ ricaun.Nuke.PackageBuilder/Components/Revit/IRevitPackageBuilder.cs
@@ -152,6 +152,7 @@ namespace ricaun.Nuke.Components
             {
                 var releaseFileName = CreateReleaseFromDirectory(BundleDirectory, projectName, projectVersion, ".bundle.zip", true);
                 Serilog.Log.Information($"Release: {releaseFileName}");
+                Serilog.Log.Information($"AppBundleTool -a \"{ReleaseDirectory / releaseFileName}\" -i");
             }
         }